Add favorites feature with UserDefaults storage
