+++ BFinder/BFinderApp.swift
@@ -9,9 +9,12 @@ import SwiftUI
 
 @main
 struct BFinderApp: App {
+    @StateObject private var favoritesManager = FavoritesManager()
+
     var body: some Scene {
         WindowGroup {
             ContentView()
+                .environmentObject(favoritesManager)
         }
     }
 }
