+++ BFinder/ContentView.swift
@@ -169,4 +169,5 @@ struct ContentView: View {
 
 #Preview {
     ContentView()
+        .environmentObject(FavoritesManager())
 }
